PR feedback, testing `Purchases.logLevel`
