+++ common/src/main/java/com/revenuecat/purchases/common/Config.kt
@@ -3,8 +3,7 @@ package com.revenuecat.purchases.common
 import com.revenuecat.purchases.LogLevel
 
 object Config {
-
-    var logLevel: LogLevel = LogLevel.debugLogsEnabled(BuildConfig.DEBUG)
+    var logLevel = LogLevel.debugLogsEnabled(BuildConfig.DEBUG)
 
     const val frameworkVersion = "5.7.0-SNAPSHOT"
 }
